Release 2 beta codebase
This is the beta codebase for Release 2. Future updates will include integration of optional platform-level identifiers for the Sierra ILS, database structure-reliant matching of usage data to EJournals stored in holdings, and a GUI (the first iteration of the GUI is introduced in the included UsageSweep.py file).

BaseLibpybr.py, Counter4pybr.py, Counter5pybr.py, and VariablesReference.py are all library files for the PySUSHI system, and requirements.txt is the dependencies file for compiling UsageSweep.py via a venv/virtualenv.